+++ packages/eslint-plugin-pf-codemods/src/rules/helpers/index.ts
@@ -7,3 +7,4 @@ export * from './JSXAttributes';
 export * from './JSXElements';
 export * from './pfPackageMatches';
 export * from './renameProps';
+export * from "./getImportDeclaration";