+++ packages/eslint-plugin-pf-codemods/src/rules/v6/chipReplaceWithLabel/chipReplaceWithLabelInput.tsx
@@ -1,0 +1,7 @@
+import { Chip } from '@patternfly/react-core/deprecated';
+
+export const ChipReplaceWithLabelInput = () => (
+  <Chip onClick={handleClick} badge={badge}>
+    This is a chip
+  </Chip>
+);