+++ packages/eslint-plugin-pf-codemods/src/rules/v6/chipReplaceWithLabel/chipReplaceWithLabelOutput.tsx
@@ -1,0 +1,8 @@
+import { Label } from '@patternfly/react-core';
+
+export const ChipReplaceWithLabelInput = () => (
+  <Label variant="outline" onClose={handleClick}>
+    This is a chip
+    {badge}
+  </Label>
+);