+++ packages/eslint-plugin-pf-codemods/src/rules/v6/formFiledGroupHeaderTitleTextObjectRenamed/formFiledGroupHeaderTitleTextObjectRenamedInput.tsx
@@ -1,4 +1,5 @@
 import { FormFiledGroupHeaderTitleTextObject } from "@patternfly/react-core";
+export { FormFiledGroupHeaderTitleTextObject as HeaderTitleObject } from "@patternfly/react-core";
 
 interface MyExtension extends FormFiledGroupHeaderTitleTextObject {}